requirements.txt
numpy
scipy
pyyaml
rich
torch
